Assume ISO paper ratio, convert signatures to A4 at 300ppi
The hardcoded size of 2480x3508 is A4 at 300 ppi.

If the given pdf file has the wrong aspect ratio, it will be stretched and
deformed to the ISO aspect ratio (sqrt(2)).

This may have a visible effect on the signature, manual cropping to ISO aspect
ratio may be required as a pre processing step.

The hardcoded pixel values (390 and 720) in signdiv.sh is the grid size
of the provided signature guide.

Fixes #8